--- a/Project.toml
+++ b/Project.toml
@@ -31,6 +31,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 SuiteSparse = "4607b0f0-06f3-5cda-b6b1-a6196a1729e9"
 TableTraits = "3783bdb8-4a98-5b6b-af9a-565f29a5fe9c"
 TerminalLoggers = "5d786b92-1e48-4d6f-9151-6b4477ca9bed"
+ThreadsX = "ac1d9e8a-700a-412c-b207-f0111f4b6c0d"
 TreeViews = "a2a6695c-b41b-5b7d-aed9-dbfdeacea5d7"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"